:bug: Fix helpers
